Create grid and slider
I created the concept for the grid along with a slider. Now I need to
connect the slider value with the squares on the grid. Right now it's
set up to work with an alert and then draw the grid.

There are many other things that will be added afterward for
functionality.